add functionalities and modifications -> restructured the code of screens int services, Controllers -> implemented rest controllers for get notes, update notes, save note and delete note -> implemented entities and getting data from db with custom queries -> refactored the code

+++ src/main/java/com/notekeepingapp/NoteKeepingAppDemo/controller/NoteController.java
@@ -5,39 +5,50 @@ import com.notekeepingapp.NoteKeepingAppDemo.helloworld.HelloWorldBean;
 import com.notekeepingapp.NoteKeepingAppDemo.model.Note;
 import com.notekeepingapp.NoteKeepingAppDemo.service.NoteService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.net.URI;
 import java.util.List;
 
+@CrossOrigin(origins = "http://localhost:3000")
 @RestController
 public class NoteController {
 
     @Autowired
     NoteService noteService;
 
-    @CrossOrigin(origins = "http://localhost:3000")
     @GetMapping("/list-notes/{name}")
-    public List<Note> showNotes(@PathVariable String name) {
-        return noteService.getNotes(name);
+    public ResponseEntity<List<Note>> showNotes(@PathVariable String name) {
+        System.out.println(name);
+        return new ResponseEntity<>(noteService.getNotes(name), HttpStatus.OK);
     }
 
-    @CrossOrigin(origins = "http://localhost:3000")
     @PostMapping("/list-notes/{name}")
-    public ResponseEntity addNote(@PathVariable String name, @RequestBody Note note) {
+    public ResponseEntity<Note> addNote(@PathVariable String name, @RequestBody Note note) {
+        System.out.println("adding note");
+        System.out.println("*************************************8");
+        System.out.println(note);
         note.setUser(name);
-        Note addedNote = noteService.addNote(note);
-        URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(addedNote.getId()).toUri();
-        return ResponseEntity.created(location).build();
+        return new ResponseEntity<>(noteService.addNote(note), HttpStatus.OK);
+    }
+
+    @PostMapping("list-notes/{name}/{id}")
+    public ResponseEntity updateNote(@PathVariable String name, @PathVariable int id, @RequestBody Note note) {
+        System.out.println(note);
+        note.setId(id);
+        note.setUser(name);
+        noteService.updateNote(note);
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     @DeleteMapping("/list-notes/{name}/{id}")
-    public void deleteNote(@PathVariable int id){
-        Note user = noteService.deleteNote(id);
-        if(user==null){
-            throw new NoteNotFoundException("id-"+id);
+    public void deleteNote(@PathVariable int id) {
+        System.out.println(id);
+        if ((noteService.isNoteExists(id))) {
+            noteService.deleteNote(id);
+        } else {
+            throw new NoteNotFoundException("id-" + id);
         }
     }
 }
